Update readme
Mention the case that inject env variables are avaialble only in the current terminal session (the on that started the container). If try to access those env vars using new session it will return the 1password item path instead of inject value.

The workaround for that can be prepending the command with `op run`:
kubectl exec \
   $(kubectl get pod -l app=app-example -o jsonpath="{.items[0].metadata.name}") \
   --container app-example1 -- /op/bin/op run -- printenv USERNAME PASSWORD

Or try to use 1password-operator to create k8s secret containing 1password item and access it from any terminal session.